Rename feature to `pull-request-hotkeys`
